HW-3  Занятие  16-17

+++ Autotests/TestAQA1/DTO/DapperTestsDTO/OrderItemsDTO.cs
@@ -11,6 +11,6 @@ namespace Autotests.TestAQA1.DTO.DapperTestsDTO
         long orderId,
         long productId,
         long quantity,
-        long unitPrice
+        decimal unitPrice
     );
 }
